requirements.txt
asgiref==3.3.1
Django==3.1.6
pytz==2021.1
sqlparse==0.4.1
django-phone-field==1.8.1
django-bootstrap-forms==0.1
django-crispy-forms==1.11.0
django-forms-bootstrap==3.1.0
mysqlclient==2.0.3
Babel==2.9.0
djangorestframework==3.12.2
requests~=2.25.1
websocket-client~=0.57.0